Refactor library
Looking to represent the relationship between
`DirectoryHandler`, `FileHandler` and a new
`Runner` object more explict, and therefore
easier to read/understand.

This resulted in a sort of re-representation
of the `Entity` object as the `Runner` object.

I have also removed `Modifier` as I am not
that this is correctly representing anything;
Living with duplicatin for now.

I suffer from adding pre-emptive optimisation
which I am looking to unlearn as a habit.